+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - chore: replace poetry with uv, make sure python3.9+ is supported and add tests for all supported python versions in github actions
+- feat: add `__call__` method to `Immutable` class to allow replacing attributes with new values, basically a shorthand for `dataclasses.replace`
 
 ## Version 1.1.1